Activator all fixed and added Anti-AFK

+++ E2Utility/Template/Template/Summoners.cpp
@@ -240,7 +240,7 @@ void Summoners::MenuLoader()
 
 void Summoners::TickLoader()
 {
-	if ((LastTimeTickCountSumm + (DWORD)Menu::Get<int>("Activator.Config.HumanizerDelay") >= GetTickCount()))
+	if ((LastTimeTickCountSumm + Menu::Get<int>("Activator.Config.HumanizerDelay") >= GetTickCount()))
 	{
 		return;
 	}
